requirements.txt
streamlit
fpdf
openai
pillow
